Update 3_reName.R
Stopped the script from deleting images of non-altered image names

+++ bin/scripts/scripts/3_reName.R
@@ -57,21 +57,27 @@ reset_names <- function(){
       aList <- list.files()
       for (aL in aList){
         setwd(aL)
-        tifList <- list.files(pattern = "tif")
-        for (tif in tifList){
-          file.remove(tif)
-        }
-        dirList <- list.files()
-        for (dir in dirList){
-          if (dir != "originals"){
-            unlink(dir, recursive = T)
+        check <- list.files()
+        if ("originals" %in% check){
+          doubleCheck <- list.files(path = "originals/")
+          if (length(doubleCheck) > 0){
+            tifList <- list.files(pattern = "tif")
+            for (tif in tifList){
+              file.remove(tif)
+            }
+            dirList <- list.files()
+            for (dir in dirList){
+              if (dir != "originals"){
+                unlink(dir, recursive = T)
+              }
+            }
+            oriFiles <- list.files(path = "originals/")
+            for (ori in oriFiles){
+              file.copy(from = paste0("originals/", ori), to = paste0("./", ori))
+            }
+            unlink("originals", recursive = T)
           }
         }
-        oriFiles <- list.files(path = "originals/")
-        for (ori in oriFiles){
-          file.copy(from = paste0("originals/", ori), to = paste0("./", ori))
-        }
-        unlink("originals", recursive = T)
         setwd("../")
       }
     }
